requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
pillow
numpy
pandas
scikit-learn
transformers
torch
sentencepiece
